Refactor Import Raw modal & mark new target as green
