+++ components/fund/permissions/MenuLeft.vue
@@ -151,14 +151,18 @@ const setSelectedTarget = (newTargetId: string) => {
 };
 
 const targetClasses = (target: Target) => {
+  const targetStatus = getTargetStatus.value(target);
   return [
-    { "permissions_menu__list_item--deleted": getTargetStatus.value(target) === EntityStatus.REMOVE },
+    { "permissions_menu__list_item--deleted": targetStatus === EntityStatus.REMOVE },
+    { "permissions_menu__list_item--new": targetStatus === EntityStatus.PENDING },
     { "permissions_menu__list_item--selected": target.id === activeTargetId.value },
   ];
 };
 const memberClasses = (memberAddress: string) => {
+  const memberStatus = getMemberStatus.value(memberAddress);
   return [
-    { "permissions_menu__list_item--deleted": getMemberStatus.value(memberAddress) === EntityStatus.REMOVE },
+    { "permissions_menu__list_item--deleted": memberStatus === EntityStatus.REMOVE },
+    { "permissions_menu__list_item--new": memberStatus === EntityStatus.PENDING },
   ];
 };
 
@@ -214,6 +218,9 @@ watchEffect(async () => {
     &--deleted {
       color: $color-disabled;
       //color: $color-error;
+    }
+    &--new {
+      color: $color-success;
     }
   }
   &__list_item_label {
